Split lib.rs in submodules and publicly export their contents
This commit refactors the lib.rs into multiple files, for better
maintainability and seperation of concerns.
