+++ src/registration.rs
@@ -7,7 +7,9 @@ use std::convert::TryInto;
 use crate::{
     apps::{App, AppBuilder},
     scopes::Scopes,
-    Data, Error, Mastodon, MastodonBuilder, Result,
+    mastodon::Mastodon,
+    mastodon::MastodonBuilder,
+    Data, Error, Result,
 };
 
 const DEFAULT_REDIRECT_URI: &str = "urn:ietf:wg:oauth:2.0:oob";
